Simplify and DRY up the indexer package
