+++ pkg/storage/utils/indexer/index/non_unique_test.go
@@ -1,4 +1,4 @@
-package disk
+package index_test
 
 import (
 	"fmt"
@@ -6,11 +6,11 @@ import (
 	"path"
 	"testing"
 
-	"github.com/cs3org/reva/pkg/storage/utils/indexer/config"
-	"github.com/cs3org/reva/pkg/storage/utils/indexer/errors"
-	"github.com/cs3org/reva/pkg/storage/utils/indexer/index"
-	"github.com/cs3org/reva/pkg/storage/utils/indexer/option"
-	. "github.com/cs3org/reva/pkg/storage/utils/indexer/test"
+	"github.com/owncloud/ocis/ocis-pkg/indexer/config"
+	"github.com/owncloud/ocis/ocis-pkg/indexer/errors"
+	"github.com/owncloud/ocis/ocis-pkg/indexer/index"
+	"github.com/owncloud/ocis/ocis-pkg/indexer/option"
+	. "github.com/owncloud/ocis/ocis-pkg/indexer/test"
 	"github.com/stretchr/testify/assert"
 )
